fix(progress-bar): [#FOR-761] fix progress bar on public side
