+++ formulaire-public/src/main/resources/public/template/containers/respond-question.html
@@ -9,7 +9,7 @@
         </section>
     </div>
 
-    <progress-bubble-bar form-element="vm.formElement" nb-form-elements="vm.nbFormElements"></progress-bubble-bar>
+    <progress-bubble-bar form-element="vm.formElement" longest-path="vm.longestPath" historic-position="vm.historicPosition"></progress-bubble-bar>
 
     <div class="respond-question">
         <div class="nine twelve-mobile">
