+++ formulaire/src/main/resources/public/ts/controllers/respond-question.ts
@@ -14,7 +14,7 @@ import {
 import {responseFileService, responseService} from "../services";
 import {FORMULAIRE_BROADCAST_EVENT, FORMULAIRE_EMIT_EVENT} from "@common/core/enums";
 import {FormElementType} from "@common/core/enums/form-element-type";
-import {UtilsUtils} from "@common/utils";
+import {FormElementUtils, UtilsUtils} from "@common/utils";
 import {Constants} from "@common/core/constants";
 
 interface ViewModel {
@@ -68,41 +68,10 @@ export const respondQuestionController = ng.controller('RespondQuestionControlle
         vm.formElement = vm.formElements.all[$scope.responsePosition - 1];
         vm.nbFormElements = vm.formElements.all.length;
         vm.historicPosition = $scope.historicPosition.length > 0 ? $scope.historicPosition : [1];
-        vm.longestPath = vm.historicPosition.length + findLongestPathInFormElement(vm.formElement.id, vm.formElements) - 1;
+        vm.longestPath = vm.historicPosition.length + FormElementUtils.findLongestPathInFormElement(vm.formElement.id, vm.formElements) - 1;
         initFormElementResponses();
         window.setTimeout(() => vm.loading = false,500);
         $scope.safeApply();
-    }
-
-
-    const findLongestPathInFormElement = (formElementId: number, formElements: FormElements): number => {
-        let currentNode: FormElement = formElements.all.find((node: FormElement) => node.id === formElementId);
-        if (!currentNode) return 1;
-        if (currentNode.isSection()) {
-            let questions: Question[] = (<Section>currentNode).questions.all;
-            let conditionalQuestions: any = questions.filter((q: Question) => q.conditional);
-            let choices: Question[] = (conditionalQuestions && conditionalQuestions.length > 0) ? conditionalQuestions.flatMap((q: Question) => q.choices.all) : null;
-            return findLongestPathInQuestionChoices(choices, currentNode, formElements);
-        } else {
-            let question: Question = <Question>currentNode;
-            let questionChoices: QuestionChoice[] = question.conditional ? question.choices.all : null;
-            return findLongestPathInQuestionChoices(questionChoices, currentNode, formElements);
-        }
-    };
-
-
-    const findLongestPathInQuestionChoices = (choices: any, currentFormElement: FormElement, formElements: FormElements): number => {
-        if (!choices || choices.length === 0) {
-            let nextElementId: number = currentFormElement.getNextFormElementId(formElements);
-            if (!nextElementId) {return 1;}
-            return findLongestPathInFormElement(nextElementId, formElements) + 1;
-        } else {
-            let tab: number[] = choices.map((choice: any) => {
-                if (!choice.next_form_element_id) return 1;
-                return findLongestPathInFormElement(choice.next_form_element_id, formElements);
-            });
-            return Math.max(...tab) + 1;
-        }
     }
 
     const initFormElementResponses = () : void => {
@@ -154,7 +123,7 @@ export const respondQuestionController = ng.controller('RespondQuestionControlle
             await saveResponses();
             vm.formElement = vm.formElements.all[prevPosition - 1];
             vm.historicPosition.pop();
-            vm.longestPath = vm.historicPosition.length + findLongestPathInFormElement(vm.formElement.id, vm.formElements) - 1;
+            vm.longestPath = vm.historicPosition.length + FormElementUtils.findLongestPathInFormElement(vm.formElement.id, vm.formElements) - 1;
             goToFormElement();
         };
         vm.isProcessing = false;
@@ -170,9 +139,8 @@ export const respondQuestionController = ng.controller('RespondQuestionControlle
         if (nextPosition && nextPosition <= vm.nbFormElements) {
             await saveResponses();
             vm.formElement = vm.formElements.all[nextPosition - 1];
-            vm.longestPath = findLongestPathInFormElement(vm.formElement.id, vm.formElements);
             vm.historicPosition.push(vm.formElement.position);
-            vm.longestPath = vm.historicPosition.length + findLongestPathInFormElement(vm.formElement.id, vm.formElements) - 1;
+            vm.longestPath = vm.historicPosition.length + FormElementUtils.findLongestPathInFormElement(vm.formElement.id, vm.formElements) - 1;
             vm.isProcessing = false;
             goToFormElement();
         }
